qwen3_tts: switch CP activations from fp32 to bf16
CP was running fp32 activations end-to-end with bf16-only ops (RoPE, weight
matmuls) in the middle, forcing ~50 TypecastDeviceOperations per 5-layer CP
block plus the fp32 SDPA cost. With bf16 activations the typecasts disappear
and SDPA runs natively bf16.

CP block kernel time:
- CP_Prefill   2239.88 → 1989.01 µs  (-11%, -251 µs, -50 ops)
- CP_Decode    2250.08 → 2001.05 µs  (-11%, -249 µs, -50 ops)
- Matmul       938 → 861 µs          (-8%, faster bf16 matmul)
- Binary ops   89 → 67 µs            (-25%)

Per-frame CP cost (1× prefill + 14× decode replays): 33.74 → 29.99 ms (-11%).

End-to-end:
- TTFT             79.4 → 66.1 ms       (-17%)
- Steady decode    56.9 → 50.2 ms/frame (-12%)
- Throughput      17.56 → 19.93 fps     (+13.5%)
- RTF (5 s audio)  0.69 → 0.61

Verification:
- Talker chained 28-layer PCC: PASSED
- CP step PCC                : PASSED
- ECAPA TTNN-vs-reference    : min 0.9997, mean 0.9998 (12 voices, unchanged)

Memory note "bf16 activations dominate when CP audio is wrong" was about
intra-CP precision when KV cache was fp32 but K/V producers were bf16 — a
*partial* fp32 path produces no benefit. Native bf16 throughout is fine, as
confirmed by PCC + ECAPA.

+++ models/demos/qwen3_tts/tt/code_predictor.py
@@ -51,7 +51,10 @@ class CodePredictor(LightweightModule):
         TILE = ttnn.TILE_LAYOUT
         ROW = ttnn.ROW_MAJOR_LAYOUT
 
-        self.act_dtype = ttnn.float32
+        # bf16 activations (was float32) — drops ~50 TypecastDeviceOperations across
+        # the 5-layer CP block (≈100 µs saving per CP_Prefill / CP_Decode replay).
+        # RoPE / KV cache / SDPA all run native bf16 now.
+        self.act_dtype = ttnn.bfloat16
         self.kcfg = ttnn.WormholeComputeKernelConfig(
             math_fidelity=ttnn.MathFidelity.HiFi4,
             math_approx_mode=False,
